Update Storyboard to Xcode 8’s Interface Builder
The set of ‘automatically created’ diffs made by this version Interface Builder trigger an /actual/, concrete change in layout behavior (which is odd - arguably even a bug - because these ‘automatically created’ IB diffs typically only apply to hidden xib metadata and are definitely not supposed to impact the the actual behavior or configuration of the layout).

The solution to this seemed to be moving the timing of the initial `reloadData` call to be performed after the view has been laid out at least once by the system.